+++ crates/biome_js_analyze/tests/specs/style/useFilenamingConvention/%validMatch.ts
@@ -1,0 +1,1 @@
+export const C: number;